[VAS-1080] refactor by adding specific bundle controller for generic APIs and added unit tests
